move redis tokenstore to subdirectory
